on ne peux plus delete un admin

+++ backend/src/controllers/userController.js
@@ -66,33 +66,58 @@ class userController {
   }
 
   static async deleteUser(req, res) {
-    const { data, error } = await crudService.destroy(User, req.params.id);
-    if (error) {
-      return res.status(404);
-    }
-    res.status(204).json({ user: data });
+      try {
+          const user = await User.findByPk(req.params.id);
+
+          if (!user) {
+              return res.status(404).json({ error: 'User not found' });
+          }
+
+          if (user.role === 'admin') {
+              return res.status(403).json({ error: 'Cannot delete admin user' });
+          }
+
+          const { data, error } = await crudService.destroy(User, req.params.id);
+          if (error) {
+              return res.status(404).json({ error: 'Error deleting user' });
+          }
+
+          res.status(204).json({ user: data });
+      } catch (error) {
+          res.status(500).json({ error: 'An error occurred' });
+      }
   }
 
   static async deleteMultiplesUsers(req, res) {
-    const { usersId } = req.body;
-    const ids = usersId.split(",");
+      const { usersId } = req.body;
+      const ids = usersId.split(",");
 
-    try {
-      const deletionPromises = ids.map(async (id) => {
-        const { data, error } = await crudService.destroy(User, id);
-        if (error) {
-          throw new Error(`User with ID ${id} not found: ${error.message}`);
-        }
-      });
+      try {
+          const users = await User.findAll({
+              where: {
+                  id: ids
+              }
+          });
 
-      await Promise.all(deletionPromises);
+          const nonAdminUsers = users.filter(user => user.role !== 'admin');
+          const nonAdminIds = nonAdminUsers.map(user => user.id);
 
-      res.sendStatus(204);
-    } catch (error) {
-      console.error("Deletion error:", error);
-      res.status(500).json({ error: "An error occurred while deleting the users" });
-    }
+          const deletionPromises = nonAdminIds.map(async (id) => {
+              const { data, error } = await crudService.destroy(User, id);
+              if (error) {
+                  throw new Error(`User with ID ${id} not found: ${error.message}`);
+              }
+          });
+
+          await Promise.all(deletionPromises);
+
+          res.sendStatus(204);
+      } catch (error) {
+          console.error("Deletion error:", error);
+          res.status(500).json({ error: "An error occurred while deleting the users" });
+      }
   }
+
 
   static async replaceUser(req, res) {
     const { data: deletedData, error: deleteError } = await crudService.destroy(User, req.params.id);
