+++ frontend/src/composables/api/user.service.ts
@@ -71,19 +71,38 @@ export const UserService = () => {
     }
 
     const deleteUser = async (id: string) => {
-        return await fetch(baseUrl + Api.user + id, {
+        const response = await fetch(`${baseUrl}${Api.user}${id}`, {
             method: 'DELETE',
-        }).then(res => res);
+        });
+        if (!response.ok) {
+            const errorData = await response.json();
+            throw new Error(errorData.error || 'Error deleting user');
+        }
+        return response;
     }
 
-    const deleteBatchUsers = async (usersId: string) => {
-        return await fetch(baseUrl + Api.user, {
+    const deleteBatchUsers = async (users: User[]) => {
+        const nonAdminUsers = users.filter(user => user.role !== 'admin');
+        const nonAdminIds = nonAdminUsers.map(user => user.id);
+        
+        if (nonAdminIds.length === 0) {
+            throw new Error('Impossible de supprimer un administrateur');
+        }
+    
+        const response = await fetch(`${baseUrl}${Api.user}`, {
             method: 'DELETE',
             headers: {
                 'Content-Type': 'application/json'
             },
-            body: JSON.stringify({usersId})
-        }).then(res => res);
+            body: JSON.stringify({ usersId: nonAdminIds.join(',') })
+        });
+    
+        if (!response.ok) {
+            const errorData = await response.json();
+            throw new Error(errorData.error || 'Une erreur est survenue');
+        }
+    
+        return response;
     }
 
     const createUser = async (bodyRequest: createUser): Promise<{sessionId: string}> => {
